update cuda code

+++ lib/utils/utils.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 
-void printMatrix(double *matrix, int m, int n) {
+void printMatrix(float *matrix, int m, int n) {
     for (int i = 0; i < m; i++) {
         for (int j = 0; j < n; j++)
             std::cout << matrix[i * m + j] << " ";
